backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
langchain>=0.1.0
langchain-community>=0.0.20
faiss-cpu>=1.13.0
pypdf>=3.17.0
python-dotenv>=1.0.0
requests>=2.31.0
pydantic>=2.0.0
sentence-transformers>=5.2.0
